transport/http: move handler into separate file
